+++ src/interfaces/cli/optimize/sqlite_file.rs
@@ -2,7 +2,7 @@ use std::io::Write;
 use std::time::Instant;
 
 pub fn command_handler_optimize_sqlite_file(sqlite_path: &str) {
-  crate::cli::require_sqlite(sqlite_path);
+  crate::interfaces::cli::require_sqlite(sqlite_path);
   print!("\x1b[1;32moptimizing\x1b[0m sqlite file...");
   let _ = std::io::stdout().flush();
   let start = Instant::now();